And now we begin using our new coordinate system (4 unit wide tiles).
git-svn-id: svn://10.0.50.55/legends/trunk@723 67207eb2-9daa-044d-9a34-b8567af74898

+++ code/src/GameSurvive/EngineLifeFormula/RoomViewer.h
@@ -50,7 +50,7 @@ public:
 	{
 		HalfSize = Real(Size) * Real::Half;
 
-		TileSize = Real(6); // Should equal RegionSize / Size
+		TileSize = Real(4); // Should equal RegionSize / Size
 		TileHalfSize = TileSize * Real::Half;
 	}
